add last bits of code
